The world and you (not really)
Adding punching and started working on a memory and emotion system.

Worked on a few more physics issues.

Trees now drop stuff.

+++ Assets/MyStuff/Scripts/ControllableHuman.cs
@@ -1,45 +1,44 @@
 using UnityEngine;
 using System.Threading.Tasks;
+using Unity.VisualScripting;
 
 public class ControllableHuman : MonoBehaviour {
 
     public static Human Instance;
+    Vector2 movement;
 
     private void Awake() {
         if (Instance != null) {
             KillPlayer();
         }
-        Instance = HumanManager.CreateRandom();
-        Debug.Log(Instance.stats);
     }
 
     public void Update() {
         if (Instance == null) return;
-        CheckMovement();
+        CheckControls();
+    }
+    public void FixedUpdate() {
+        if (Instance == null) return;
+        SetPosition();
     }
 
     void CheckControls() {
-        Vector2 movement = Vector2.zero;
-
+        movement = Vector2.zero;
         if (Input.GetKey(KeyCode.W)) movement.y++;
         if (Input.GetKey(KeyCode.S)) movement.y--;
 
         if (Input.GetKey(KeyCode.D)) movement.x++;
         if (Input.GetKey(KeyCode.A)) movement.x--;
 
-        Vector2 pos = (Vector2)Instance.transform.position;
-        pos += movement.normalized * Instance.stats.speed * Time.deltaTime;
-        Instance.gameObject.GetComponent<Rigidbody2D>().MovePosition(pos);
+        if (Input.GetMouseButtonDown(0)) Instance.Punch(1f);
     }
-    async Task CheckMovement() {
-        if (!Input.GetMouseButtonDown(1)) return;
-
-        Vector2 pos = CameraManager.GetMousePosition;
-        await Instance.WalkTowards(pos);
+    void SetPosition() {
+        Instance.Move(movement);
+        Instance.LookAtPos(CameraManager.GetMousePosition);
     }
-
 
     public static void KillPlayer() {
+        if (Instance == null) return;
         Instance.Kill();
         Instance = null;
         GameManager.isPlaying = false;
